533 Move package optionsets into their own file
