+++ client/app/helpers/optionset.js
@@ -15,7 +15,7 @@ import {
   PACKAGE_STATUS_OPTIONSET,
   PACKAGE_VISIBILITY_OPTIONSET,
   PACKAGE_TYPE_OPTIONSET,
-} from '../models/package';
+} from '../optionsets/package';
 import {
   DCPHASPROJECTCHANGEDSINCESUBMISSIONOFTHEPAS_OPTIONSET,
   DCPCONSTRUCTIONPHASING_OPTIONSET,
